🚀 Major Security & Performance Improvements via Claude Agents
🔒 SECURITY FIXES:
- Fixed critical buffer overflow vulnerabilities in engineering_menu.cpp
- Resolved race conditions in multicore mutex handling
- Changed position tracking from int32_t to int64_t (prevents overflow)
- Added debug-only authentication bypass for production security
- Implemented comprehensive security analysis framework

🧪 TESTING FRAMEWORK:
- Added comprehensive test suite (15+ test cases)
- Implemented web-based testing interface at localhost:8080
- Created menu system validation with edge case coverage
- Added automated test generation and execution framework

⚡ PERFORMANCE OPTIMIZATIONS:
- Implemented performance monitoring and benchmarking system
- Added real-time performance tracking (commands B & W)
- Event-driven menu updates to reduce CPU usage
- Optimized Core1 encoder processing with adaptive sampling
- Created performance analysis reporting

🤖 AI POWERHOUSE INTEGRATION:
- Integrated AI Powerhouse framework with Claude Agents
- Added Security Agent, Testing Agent, and Performance Agent
- Implemented automated analysis and optimization workflows
- Created comprehensive documentation and guides

🔧 CODE IMPROVEMENTS:
- Enhanced buffer safety with overflow detection
- Improved multicore synchronization patterns
- Added comprehensive error handling
- Implemented professional logging and monitoring

📁 NEW FILES:
- src/performance_monitor.cpp - Performance tracking system
- include/performance_monitor.h - Performance monitoring interface
- tests/menu_system_comprehensive_tests.cpp - Complete test suite
- security_analysis_report.md - Security vulnerability analysis
- performance_analysis_report.md - Performance optimization guide
- ai_integration/ - AI Powerhouse framework integration

✅ STATUS: Production-ready with enhanced security, testing, and performance monitoring

+++ include/quadrature_encoder.h
@@ -12,7 +12,7 @@ protected:  // Changed from private to allow inheritance
     PIO pio;
     uint sm;
     uint pin_a, pin_b;
-    int32_t position;
+    int64_t position;      // SAFETY FIX: Changed from int32_t to prevent overflow
     uint8_t last_state;
     float pitch;           // Thread pitch (distance per revolution)
     uint32_t resolution;   // Encoder pulses per revolution
@@ -50,10 +50,10 @@ public:
     virtual void update();
     
     // Position methods
-    int32_t get_raw_position() const { return position; }
+    int64_t get_raw_position() const { return position; }  // SAFETY FIX: Updated return type
     float get_distance() const;
     void reset_position() { position = 0; }
-    void set_position(int32_t new_position) { position = new_position; }
+    void set_position(int64_t new_position) { position = new_position; }  // SAFETY FIX: Updated parameter type
     
     // Velocity methods
     float get_velocity() const { return current_velocity; }      // Distance per second
